requirements.txt
Django>=5.2
gunicorn
python-decouple
psycopg[binary]
dj-database-url
requests
whitenoise
django-allauth[socialaccount]
django-allauth-ui
django-widget-tweaks
slippers
stripe
djangorestframework
django-filter
markdown
djangorestframework-simplejwt
pyyaml
django-cors-headers